API

This is the documentation of the Arianna module's functions, types and structures.

API Reference

Index

Arianna.Arianna

Arianna.Action

Arianna.AriannaAlgorithm

Arianna.AriannaSystem

Arianna.DAT

Arianna.Format

Arianna.Metropolis

Arianna.Metropolis

Arianna.Move

Arianna.Move

Arianna.Policy

Arianna.PrintTimeSteps

Arianna.Simulation

Arianna.Simulation

Arianna.StoreBackups

Arianna.StoreCallbacks

Arianna.StoreLastFrames

Arianna.StoreParameters

Arianna.StoreParameters

Arianna.StoreTrajectories

Arianna.TXT

Arianna.build_schedule

Arianna.build_schedule

Arianna.build_schedule

Arianna.callback_acceptance

Arianna.delta_log_target_density

Arianna.finalise

Arianna.finalise

Arianna.initialise

Arianna.initialise

Arianna.invert_action!

Arianna.log_proposal_density

Arianna.make_step!

Arianna.make_step!

Arianna.mc_step!

Arianna.mc_sweep!

Arianna.perform_action!

Arianna.perform_action_cached!

Arianna.raise_error

Arianna.run!

Arianna.sample_action!

Arianna.store_trajectory

Arianna.unnormalised_log_target_density

Arianna.write_algorithm

Arianna.write_algorithm

Arianna.write_parameters

Types and Functions

Arianna.Arianna — Module

module Arianna

Arianna is a flexible and extensible framework for Monte Carlo simulations. Instead of acting as a black-box simulator, it provides a modular structure where users define their own system and Monte Carlo "moves".

Arianna.Action — Type

abstract type Action

Abstract type representing Monte Carlo actions/moves that can be performed on a system.

Concrete subtypes must implement:

sample_action!(action, policy, parameters, system, rng): Sample a new action from the policy

perform_action!(system, action): Apply the action to modify the system state

invert_action!(action, system): Invert/reverse the action

log_proposal_density(action, policy, parameters, system): Log probability density of proposing this action

Optional methods:

perform_action_cached!(system, action): Optimized version of perform_action! that can reuse cached values

Arianna.AriannaAlgorithm — Type

abstract type AriannaAlgorithm

Abstract type for Simulation algorithms.

Arianna.AriannaSystem — Type

abstract type AriannaSystem

Abstract type representing a system that can be simulated using methods defined in the Arianna module.

Arianna.DAT — Type

DAT <: Format

Format type for data (.dat) file output.

Arianna.Format — Type

Format

Abstract type for output file formats.

Arianna.Metropolis — Type

Metropolis{P,R<:AbstractRNG,C<:Function} <: AriannaAlgorithm

A struct representing a Metropolis Monte Carlo algorithm.

Fields

pools::Vector{P}: Vector of independent pools (one for each system)

sweepstep::Int: Number of Monte Carlo steps per sweep

seed::Int: Random number seed

rngs::Vector{R}: Vector of random number generators

parallel::Bool: Flag to parallelise over different threads

collecter::C: Transducer to collect results from parallelised loops

Type Parameters

P: Type of the pool

R: Type of the random number generator

C: Type of the transducer

Arianna.Metropolis — Method

Metropolis(chains; pool=missing, sweepstep=1, seed=1, R=Xoshiro, parallel=false, extras...)

Create a new Metropolis instance.

Arguments

chains: Vector of chains to run the algorithm on

pool: Pool of moves to perform sweeps over

sweepstep: Number of Monte Carlo steps per sweep

seed: Random number seed

R: Type of the random number generator

parallel: Flag to parallelise over different threads

extras: Additional keyword arguments

Returns

algorithm: Metropolis instance

Arianna.Move — Type

Move{A<:Action,P<:Policy,V<:AbstractArray,T<:AbstractFloat}

A struct representing a Monte Carlo move with an associated action, policy, and parameters.

Fields

action::A: The action to be performed in the move

policy::P: The policy used to propose actions

parameters::V: Parameters of the policy

weight::T: Weight/probability of selecting this move in a sweep

total_calls::Int: Total number of times this move has been attempted

accepted_calls::Int: Number of times this move has been accepted

Type Parameters

A: Type of the action

P: Type of the policy

V: Type of the parameter array

T: Type of the weight (floating point)

Arianna.Move — Method

Move(action, policy, parameters, weight)

Create a new Move instance.

Arguments

action: The action to be performed in the move

policy: The policy used to propose actions

parameters: Parameters of the policy

weight: Weight/probability of selecting this move in a sweep

Arianna.Policy — Type

abstract type Policy

Abstract type representing proposal policies for Monte Carlo actions.

A Policy defines how actions are sampled and their proposal probabilities are calculated. Concrete subtypes work together with specific Action types to implement the proposal mechanism.

Arianna.PrintTimeSteps — Type

PrintTimeSteps <: AriannaAlgorithm

Algorithm to display a progress bar and current timestep during simulation.

Arianna.Simulation — Type

mutable struct Simulation{S, A, VS}

A structure representing a Monte Carlo simulation.

Fields

chains::Vector{S}: Vector of independent Arianna systems.

algorithms::A: List of algorithms.

steps::Int: Number of MC sweeps.

t::Int: Current time step.

schedulers::VS: List of schedulers (one for each algorithm).

counters::Vector{Int}: Counters for the schedulers (one for each algorithm).

path::String: Simulation path.

verbose::Bool: Flag for verbose output.

Arianna.Simulation — Method

Simulation(chains, algorithm_list, steps; path="data", verbose=false)

Create a new Simulation instance from a list of algorithm constructors.

Arguments

chains: Vector of independent Arianna systems.

algorithm_list: List of algorithm constructors.

steps: Number of MC sweeps.

path="data": Simulation path.

verbose=false: Flag for verbose output.

Arianna.StoreBackups — Type

StoreBackups <: AriannaAlgorithm

Algorithm to create backup files of system states during simulation.

Fields

dirs::Vector{String}: Directories for storing backup files

fmt::Format: Format type for output files

store_first::Bool: Whether to store backups at initialization

store_last::Bool: Whether to store backups at finalization

Arianna.StoreCallbacks — Type

StoreCallbacks{V} <: AriannaAlgorithm

Algorithm to store callback values during simulation.

Fields

callbacks::V: Vector of callback functions to evaluate

paths::Vector{String}: Paths to output files for each callback

files::Vector{IOStream}: File handles for writing callback values

store_first::Bool: Whether to store callback values at initialization

store_last::Bool: Whether to store callback values at finalization

Constructor

StoreCallbacks(callbacks::V, path; store_first::Bool=true, store_last::Bool=false)

Create a new StoreCallbacks instance.

Arguments

callbacks::V: Vector of callback functions

path: Base path for output files

store_first=true: Store values at initialization

store_last=false: Store values at finalization

Arianna.StoreLastFrames — Type

StoreLastFrames <: AriannaAlgorithm

Algorithm to store the final state of each system at the end of simulation.

Fields

paths::Vector{String}: Paths to output files

fmt::Format: Format type for output files

Arianna.StoreParameters — Type

StoreParameters{V<:AbstractArray} <: AriannaAlgorithm

A struct representing a parameter store.

Fields

paths::Vector{String}: Vector of paths to the parameter files

files::Vector{IOStream}: Vector of open file streams to the parameter files

parameters_list::V: List of parameters to store

store_first::Bool: Flag to store the parameters at the first step

store_last::Bool: Flag to store the parameters at the last step

Type Parameters

V: Type of the parameter array

Arianna.StoreParameters — Method

StoreParameters(chains; dependencies=missing, path=missing, ids=missing, store_first=true, store_last=false, extras...)

Create a new StoreParameters instance.

Arguments

chains: Vector of chains to store the parameters of

dependencies: Dependencies of the parameter store

path: Path to the parameter files

ids: IDs of the parameters to store

store_first: Flag to store the parameters at the first step

store_last: Flag to store the parameters at the last step

extras: Additional keyword arguments

Returns

algorithm: StoreParameters instance

Arianna.StoreTrajectories — Type

StoreTrajectories{F<:Format} <: AriannaAlgorithm

Algorithm to store system trajectories during simulation.

Fields

paths::Vector{String}: Paths to output files

files::Vector{IOStream}: File handles for writing trajectories

fmt::F: Format type for output files

store_first::Bool: Whether to store trajectories at initialization

store_last::Bool: Whether to store trajectories at finalization

Arianna.TXT — Type

TXT <: Format

Format type for text (.txt) file output.

Arianna.build_schedule — Method

build_schedule(steps::Int, burn::Int, base::AbstractFloat)

Create a vector of timestep from burn to steps log-spaced with base base.

Arianna.build_schedule — Method

build_schedule(steps::Int, burn::Int, Δt::Int)

Create a vector of timestep from burn to steps at intervals Δt.

Arianna.build_schedule — Method

build_schedule(steps::Int, burn::Int, block::Vector{Int})

Create a vector of timestep from burn to steps with repeated blocks specified by block.

Arianna.callback_acceptance — Method

callback_acceptance(simulation)

Calculate the mean acceptance rate of the Metropolis algorithm.

Arguments

simulation: Simulation to calculate the acceptance rate of

Arianna.delta_log_target_density — Method

delta_log_target_density(x₁, x₂, system::AriannaSystem)

Calculate the change in log target density between two states.

Arguments

x₁: Initial state

x₂: Final state

system: System object

Arianna.finalise — Method

finalise(::AriannaAlgorithm, ::Simulation)

Finalise the algorithm for the given simulation.

Arianna.finalise — Method

store_lastframe(io, system, t, fmt::Format)

Store the final state of the system at time t to the given IO stream in the specified format.

Arianna.initialise — Method

initialise(::AriannaAlgorithm, ::Simulation)

Initialise the algorithm for the given simulation.

Arianna.initialise — Method

store_backup(io, system, t, fmt::Format)

Store a backup of the system state at time t to the given IO stream in the specified format.

Arianna.invert_action! — Method

invert_action!(action::Action, system::AriannaSystem)

Invert/reverse an action.

Arguments

action: Action to invert

system: System the action was applied to

Arianna.log_proposal_density — Method

log_proposal_density(action, policy, parameters, system::AriannaSystem)

Calculate the log probability density of proposing the given action.

Arguments

action: Proposed action

policy: Policy used for proposal

parameters: Parameters of the policy

system: System the action is applied to

Arianna.make_step! — Method

make_step!(::Simulation, ::AriannaAlgorithm)

Perform a single step of the algorithm in the simulation.

Arianna.make_step! — Method

make_step!(simulation::Simulation, algorithm::Metropolis)

Perform a single step of the Metropolis algorithm.

Arguments

simulation: Simulation to perform the step on

algorithm: Metropolis instance

Arianna.mc_step! — Method

mc_step!(system::AriannaSystem, action::Action, policy::Policy, parameters::AbstractArray{T}, rng) where {T<:AbstractFloat}

Perform a single Monte Carlo step.

Arguments

system: System to modify

action: Action to perform

policy: Policy used for proposal

parameters: Parameters of the policy

rng: Random number generator

Arianna.mc_sweep! — Method

mc_sweep!(system::AriannaSystem, pool, rng; mc_steps=1)

Perform a Monte Carlo sweep over a pool of moves.

Arguments

system: System to modify

pool: Pool of moves to perform sweeps over

rng: Random number generator

mc_steps: Number of Monte Carlo steps per sweep

Arianna.perform_action! — Method

perform_action!(system::AriannaSystem, action::Action)

Apply the action to modify the system state.

Arguments

system: System to modify

action: Action to perform

Returns

A tuple of (x₁, x₂) containing the old and new states

Arianna.perform_action_cached! — Method

perform_action_cached!(system, action::Action)

Optimized version of perform_action! that can reuse cached values.

Arguments

system: System to modify

action: Action to perform

Arianna.raise_error — Method

raise_error(s)

Helper function to raise errors for unimplemented required methods.

Arguments

s: String describing the missing method implementation

Arianna.run! — Method

run!(simulation::Simulation)

Run the Monte Carlo simulation.

Arguments

simulation::Simulation: The simulation instance to run.

Arianna.sample_action! — Method

sample_action!(action::Action, policy::Policy, parameters, system::AriannaSystem, rng)

Sample a new action from the policy.

Arguments

action: Action to be sampled

policy: Policy to sample from

parameters: Parameters of the policy

system: System the action will be applied to

rng: Random number generator

Arianna.store_trajectory — Method

store_trajectory(io, system::AriannaSystem, t, fmt::Format)

Store the system trajectory at time t to the given IO stream in the specified format.

Arianna.unnormalised_log_target_density — Method

unnormalised_log_target_density(x, system)

Calculate the unnormalized log probability density of a system state.

Arguments

x: System state

system: System object

Arianna.write_algorithm — Method

write_algorithm(io, algorithm::AriannaAlgorithm, scheduler)

Write a summary of the algorithm on the given IO stream.

Arianna.write_algorithm — Method

write_algorithm(io, algorithm::Metropolis, scheduler)

Write the algorithm to a string.

Arguments

io: IO stream to write to

algorithm: Algorithm to write

scheduler: Scheduler to write

Arianna.write_parameters — Method

write_parameters(::Policy, parameters)

Write the parameters of a policy to a string.

Arguments

policy: Policy to write the parameters of

parameters: Parameters of the policy

Arianna.jl

```@autodocs